+++ Blogs.Application/Services/Implementations/CommentService.cs
@@ -12,6 +12,7 @@ using Ganss.Xss;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -23,95 +24,107 @@ namespace Blogs.Application.Services.Implementations
         private readonly IMapper _mapper;
         private readonly AbstractValidator<CreateCommentDTO> _createValidator;
         private readonly AbstractValidator<UpdateCommentDTO> _updateValidator;
-        private readonly AbstractValidator<UserDTO> _userValidator;
 
         public CommentService(IUnitOfWork unitOfWork, 
             IMapper mapper, 
             AbstractValidator<CreateCommentDTO> createValidator, 
-            AbstractValidator<UpdateCommentDTO> updateValidator, 
-            AbstractValidator<UserDTO> userValidator)
+            AbstractValidator<UpdateCommentDTO> updateValidator)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
             _createValidator = createValidator;
             _updateValidator = updateValidator;
-            _userValidator = userValidator;
         }
 
-        public async Task<CommentResultDTO> CreateAsync(CreateCommentDTO comment, UserDTO user, CancellationToken cancellationToken = default)
+        public async Task<CommentResultDTO> CreateAsync(CreateCommentDTO comment, Guid userId, CancellationToken cancellationToken = default)
         {
             if (comment == null)
                 throw new ArgumentNullException(nameof(comment));
 
-            if (user == null)
-                throw new ArgumentNullException(nameof(user));
+            if (userId == null)
+                throw new ArgumentNullException(nameof(userId));
 
             var validation = await _createValidator.ValidateAsync(comment);
 
             if (!validation.IsValid)
                 throw new BusinessException(string.Join("\n", validation.Errors));
 
-            var userValidation = await _userValidator.ValidateAsync(user);
+            if (!_unitOfWork.UserRepository.Exists(userId))
+                throw new BusinessException("User does not exist");
 
-            if (!userValidation.IsValid)
-                throw new BusinessException(string.Join("\n", userValidation.Errors));
+            if (!_unitOfWork.BlogPostsRepository.Exists(comment.BlogPostId))
+                throw new BusinessException("The blog does not exist");
 
             var commentModel = _mapper.Map<Comment>(comment);
             commentModel.Id = Guid.NewGuid();
             commentModel.CreationDate = DateTime.Now;
-            commentModel.CreationUser = user.Id;
+            commentModel.CreationUser = userId;
             commentModel.UpdateDate = DateTime.Now;
-            commentModel.UpdateUser = user.Id;
+            commentModel.UpdateUser = userId;
 
             _unitOfWork.CommentRepository.Insert(commentModel);
             _unitOfWork.Commit();
 
-            return _mapper.Map<CommentResultDTO>(commentModel);
+            var result = _mapper.Map<CommentResultDTO>(commentModel);
+
+            return result;
         }
 
-        public async Task DeleteAsync(Guid id, CancellationToken cancellationToken = default)
+        public async Task DeleteAsync(Guid id, Guid userId, CancellationToken cancellationToken = default)
         {
+            var comment = _unitOfWork.CommentRepository.Get(id);
+
+            if (comment.CreationUser != userId)
+                throw new BusinessException("You can't delete this comment, you are not its author");
+
             _unitOfWork.CommentRepository.Delete(id);
             _unitOfWork.Commit();
         }
 
-        public async Task<IEnumerable<CommentResultDTO>> GetAllAsync(CancellationToken cancellationToken = default)
+        public async Task<IEnumerable<CommentResultDTO>> GetBlogPostComments(Guid blogPostId, 
+            CancellationToken cancellationToken = default)
         {
-            return _unitOfWork.CommentRepository.GetAll().Select(comment => _mapper.Map<CommentResultDTO>(comment));
+            return _unitOfWork.CommentRepository.GetAllByBlogPost(blogPostId).Select(comment => _mapper.Map<CommentResultDTO>(comment));
         }
 
-        public async Task<CommentResultDTO> GetAsync(Guid id, CancellationToken cancellationToken = default)
+        public async Task<CommentResultDTO> GetCommentOfBlogPost(Guid blogPostId, Guid id, CancellationToken cancellationToken = default)
         {
-            return _mapper.Map<CommentResultDTO>(_unitOfWork.CommentRepository.Get(id));
+            var comment = _unitOfWork.CommentRepository.GetByBlogPost(blogPostId, id);
+            var result = _mapper.Map<CommentResultDTO>(comment);
+
+            return result;
         }
 
-        public async Task<CommentResultDTO> UpdateAsync(UpdateCommentDTO comment, UserDTO user, CancellationToken cancellationToken = default)
+        public async Task<CommentResultDTO> UpdateAsync(UpdateCommentDTO comment, Guid userId, CancellationToken cancellationToken = default)
         {
             if (comment == null)
                 throw new ArgumentNullException(nameof(comment));
 
-            if (user == null)
-                throw new ArgumentNullException(nameof(user));
+            if (userId == null)
+                throw new ArgumentNullException(nameof(userId));
 
             var validation = await _updateValidator.ValidateAsync(comment);
 
             if (!validation.IsValid)
                 throw new BusinessException(string.Join("\n", validation.Errors));
 
-            var userValidation = await _userValidator.ValidateAsync(user);
-
-            if (!userValidation.IsValid)
-                throw new BusinessException(string.Join("\n", userValidation.Errors));
+            if (!_unitOfWork.UserRepository.Exists(userId))
+                throw new BusinessException("User does not exist");
 
             var commentToUpdate = _unitOfWork.CommentRepository.Get(comment.Id);
+
+            if (commentToUpdate.CreationUser != userId)
+                throw new BusinessException("You can't edit this comment, you are not its author ");
+
             commentToUpdate.UpdateDate = DateTime.Now;
-            commentToUpdate.UpdateUser = user.Id;
+            commentToUpdate.UpdateUser = userId;
 
             _unitOfWork.CommentRepository.Update(_mapper.Map(comment, commentToUpdate));
-
             _unitOfWork.Commit();
 
-            return _mapper.Map<CommentResultDTO>(comment);
+            var result = _mapper.Map<CommentResultDTO>(commentToUpdate);
+
+            return result;
         }
     }
 }
